Move host service classes to their own files
